+++ plom/db/db_upload.py
@@ -167,7 +167,7 @@ def replaceMissingTestPage(
     # we can actually just call uploadTPage - we just need to set the bundle_name and bundle_order.
     # hw is different because we need to verify no hw pages present already.
 
-    bref = Bundle.get_or_none(name="replacements")
+    bref = Bundle.get_or_none(name="__replacements__system__")
     if bref is None:
         return [False, "bundleError", f'Cannot find bundle "replacements"']
 
@@ -185,7 +185,7 @@ def replaceMissingTestPage(
         original_name,
         file_name,
         md5,
-        "replacements",
+        "__replacements__system__",
         bundle_order,
     )
     if rval[0]:  # success - so trigger an update.
@@ -379,7 +379,7 @@ def replaceMissingHWQuestion(self, sid, question, original_name, file_name, md5)
     # todo = merge this somehow with uploadHWPage? - most of function is sanity checks.
 
     order = 1
-    bundle_name = "replacements"
+    bundle_name = "__replacements__system__"
 
     iref = IDGroup.get_or_none(student_id=sid)
     if iref is None:
@@ -727,7 +727,8 @@ def updateQGroup(self, qref):
     # otherwise ready.
     scan_list = [p.scanned for p in gref.tpages]  # list never zero length.
     if True in scan_list:  # some tpages scanned.
-        if False in scan_list:  # some tpages unscanned - definitely not ready to go.
+        # some tpages unscanned - definitely not ready to go.
+        if False in scan_list:
             log.info("Group {} is only half-scanned - not ready".format(gref.gid))
             return False
         else:
@@ -929,3 +930,27 @@ def removeAllScannedPages(self, test_number):
         tref.save()
     self.updateTestAfterUpload(tref)
     return [True, "Test {} wiped clean".format(test_number)]
+
+
+### some bundle related stuff
+def listBundles(self):
+    """Returns a list of bundles in the database
+
+    Args: None
+
+    Returns:
+        List (dict). One for each bundle. Each dict contains three
+            key-value pairs: "name", "md5sum" and "numberOfPages".
+            If no bundles in the system, then it returns an empty list.
+    """
+
+    bundle_info = []
+    for bref in Bundle.select():
+        bundle_info.append(
+            {
+                "name": bref.name,
+                "md5sum": bref.md5sum,
+                "numberOfPages": len(bref.images),
+            }
+        )
+    return bundle_info