+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 * Various CI cleanups that should help external contributors.
-
+* plom-scan and plom-hwscan can now list bundles.
 
 
 ## [0.7.1] - 2021-09-23